fix(share): remove all instances of react-share
